Moved functions to datalog module
