+++ app/src/datalog_module.h
@@ -1,0 +1,1 @@
+void hpi_datalog_start_session(void);
